moving to azure auth package
